Fix multiple events in a single read (#19)
A read from the inotify descriptor may return multiple events if there
is enough space in the read buffer.  Since our read buffer includes
space for a pathname of length 4096, a single read can return up to 257
events, of which only the first one is processed by ReadSingleEvent and
all the rest is discarded.

Implement @phemmer's suggestion to preserve the read buffer across calls
by moving it into the Watcher struct.  ReadSingleEvent now reads events
from the buffer until reaching the end of the used space, and only then
it reads from the inotify descriptor to refill the buffer.